((Docker mode - compile) 避免重复下载
使用flag文件，已经下载时将会跳过下载步骤，除非移除相关flag文件

+++ builder-in-docker/openwrt-data/03-compile.sh
@@ -8,15 +8,18 @@ export GITHUB_WORKSPACE="$(pwd)"
 cd openwrt
 export OPENWRT_PATH="$(pwd)"
 
-# { Download DL Package(下载DL软件包) }
-cd "$OPENWRT_PATH"
+# { Download DL Package(下载DL软件包) }cd "$OPENWRT_PATH"
 make defconfig
-make download -j8
-find dl -size -1024c -exec ls -l {} \;
-find dl -size -1024c -exec rm -f {} \;
+if ! [[ -f "$OPENWRT_PATH/flags-downloaded-packages" ]]
+then
+    make download -j8
+    find dl -size -1024c -exec ls -l {} \;
+    find dl -size -1024c -exec rm -f {} \;
+    touch "$OPENWRT_PATH/flags-downloaded-packages"
+fi
 
 # { Compile Firmware (多线程尝试), Compile Firmware Single Thread with Verbosity (单线程详细模式) }
-cd $OPENWRT_PATH
+cd "$OPENWRT_PATH"
 echo -e "$(nproc) thread compile"
 # 记录开始编译时的时间戳
 export TIMESTAMP_START_COMPILE=$(printf "%(%s)T")
